requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
pytorch-lightning>=2.1.0

# Medical Imaging
monai>=1.3.0
pydicom>=2.4.3
SimpleITK>=2.3.0

# Computer Vision & Image Processing
opencv-python>=4.8.1.78
Pillow>=10.1.0
albumentations>=1.3.1
scikit-image>=0.22.0

# Explainability & Visualization
grad-cam>=1.4.8
pytorch-grad-cam>=1.4.8
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Data Science & ML
numpy>=1.24.3
pandas>=2.1.4
scikit-learn>=1.3.2
scipy>=1.11.4

# Web Framework & API
streamlit>=1.29.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Report Generation
reportlab>=4.0.7
fpdf2>=2.7.6
Jinja2>=3.1.2

# Utilities
PyYAML>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1
loguru>=0.7.2
colorama>=0.4.6

# Model Export & Optimization
onnx>=1.15.0
onnxruntime>=1.16.3
onnxsim>=0.4.35

# Testing & Quality
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.7.1

# Metrics & Monitoring
tensorboard>=2.15.1
wandb>=0.16.1

# Async & Concurrency
aiofiles>=23.2.1
httpx>=0.25.2

# Image Quality Assessment
pyiqa>=0.1.10
